fix: address PR review comments for #1555
- Move all allpads-r9 tests from input_tests.rs to miscellaneous_tests.rs
- Keep input_tests.rs focused on input-suite TODO placeholders
- Update stale allpads.nes expect messages to allpads218.nes
